+++ src/main/kotlin/com/aaronhowser1/pitchperfect/event/ModEvents.kt
@@ -1,14 +1,20 @@
 package com.aaronhowser1.pitchperfect.event
 
 import com.aaronhowser1.pitchperfect.PitchPerfect
+import com.aaronhowser1.pitchperfect.config.ServerConfig
 import com.aaronhowser1.pitchperfect.enchantment.AndHisMusicWasElectricEnchantment
 import com.aaronhowser1.pitchperfect.item.InstrumentItem
+import com.aaronhowser1.pitchperfect.item.ModItems
+import com.aaronhowser1.pitchperfect.utils.CommonUtils.isMonster
+import net.minecraft.world.InteractionHand
 import net.minecraft.world.entity.LivingEntity
 import net.minecraftforge.event.TickEvent
 import net.minecraftforge.event.TickEvent.ServerTickEvent
 import net.minecraftforge.event.entity.living.LivingHurtEvent
+import net.minecraftforge.event.entity.living.LivingSpawnEvent
 import net.minecraftforge.eventbus.api.SubscribeEvent
 import net.minecraftforge.fml.common.Mod
+import kotlin.random.Random
 
 @Suppress("unused")
 @Mod.EventBusSubscriber(modid = PitchPerfect.MOD_ID)
@@ -26,6 +32,18 @@ object ModEvents {
 
     }
 
+    @SubscribeEvent
+    fun onLivingSpawn(event: LivingSpawnEvent) {
+        val entity = event.entity
+        if (
+            !entity.isMonster() ||
+            !entity.mainHandItem.isEmpty ||
+            ServerConfig.MOB_SPAWNS_WITH_INSTRUMENT_CHANCE.get() < Random.nextFloat()
+        ) return
+
+        val randomInstrument = ModItems.ITEM_REGISTRY.entries.random().get()
+        entity.setItemInHand(InteractionHand.MAIN_HAND, randomInstrument.defaultInstance)
+    }
 
     @SubscribeEvent
     fun serverTick(event: ServerTickEvent) {